Most everything is done.
Have a non working copy of the Pediatric Airways example.
The only thing left that I can think of right now is to add a module that can manipulate the data.
Fixed the problem with the weighted function boxplot.

I could try to add functionality for a 3D surface boxplotbut that may be difficult.